Move ClusterAssociated to the clusters module
This patch fixes an error due to import loop
