+++ ovirtlib4/templates.py
@@ -2,7 +2,8 @@
 
 import ovirtsdk4.types as types
 
-from .system_service import CollectionService, CollectionEntity, ClusterAssociated
+from .clusters import ClusterAssociated
+from .system_service import CollectionService, CollectionEntity
 
 
 class Templates(CollectionService):
