+++ ovirtlib4/vms.py
@@ -1,10 +1,12 @@
 # -*- coding: utf-8 -*-
 
 import ipaddress
+
 import ovirtsdk4.types as types
 
 from . import defaults, hosts, vnic_profiles
-from .system_service import CollectionService, CollectionEntity, ClusterAssociated
+from .clusters import ClusterAssociated
+from .system_service import CollectionService, CollectionEntity
 
 
 class Vms(CollectionService):
